chore(routes): clean up and improve route organization
