+++ src/controllers/doctorController.js
@@ -51,7 +51,6 @@ const registerDoctor = async (req, res) => {
       },
       req.user?.id
     );
-    console.log(esp.id);
     
     return res.status(201).json({
       message: "Doctor registrado correctamente",
@@ -68,6 +67,66 @@ const registerDoctor = async (req, res) => {
     return res
       .status(500)
       .json({ message: error.message || "Error interno del servidor" });
+  }
+};
+
+const getDoctorsBySpecialty = async (req, res) => {
+  try {
+    const { specialty, page = 1 } = req.query;
+    if (!specialty) return res.status(400).json({ message: "Debe especificar 'specialty'" });
+
+    const limit = 20;
+    const skip = (parseInt(page, 10) - 1) * limit;
+
+    const especializaciones = await prisma.especializacion.findMany({
+      where: { nombre: { equals: specialty, mode: "insensitive" } },
+      select: { id: true, nombre: true, departamentoId: true },
+    });
+
+    if (!especializaciones || especializaciones.length === 0) {
+      return res.status(404).json({ message: `No se encontró la especialidad '${specialty}'` });
+    }
+
+    const espIds = especializaciones.map(e => e.id);
+
+    const total = await prisma.users.count({
+      where: {
+        role: "MEDICO",
+        status: "ACTIVE",
+        especializacionId: { in: espIds },
+      },
+    });
+
+    const doctors = await prisma.users.findMany({
+      where: {
+        role: "MEDICO",
+        status: "ACTIVE",
+        especializacionId: { in: espIds },
+      },
+      skip,
+      take: limit,
+      orderBy: { createdAt: "desc" },
+      select: {
+        id: true,
+        fullname: true,
+        email: true,
+        role: true,
+        status: true,
+        createdAt: true,
+        updatedAt: true,
+        especializacion: { select: { nombre: true } }
+      },
+    });
+
+    return res.status(200).json({
+      page: parseInt(page, 10),
+      total,
+      totalPages: Math.ceil(total / limit),
+      doctors,
+    });
+  } catch (err) {
+    console.error("Error en getDoctorsBySpecialty:", err);
+    return res.status(500).json({ message: "Error interno del servidor" });
   }
 };
 
@@ -213,6 +272,7 @@ const toggleDoctorStatus = async (req, res) => {
   }
 };
 
-module.exports = { registerDoctor, getDoctorById, updateDoctor, toggleDoctorStatus };
+
+module.exports = { registerDoctor, getDoctorById, updateDoctor, toggleDoctorStatus, getDoctorsBySpecialty };
