+++ src/router/userRoutes.js
@@ -3,10 +3,11 @@ const router = express.Router();
 
 const doctorRoutes = require("./doctorRoutes");
 const nurseRoutes = require("./nurseRoutes");
-const { getUserById } = require("../controllers/UserController");
+const { getUserById, getUsersByRole } = require("../controllers/UserController");
+const authMiddleware = require("../middleware/authMiddleware");
 
+router.get("/by-role" , authMiddleware, getUsersByRole);
 router.get("/:userId", getUserById);
-
 router.use("/doctors", doctorRoutes);
 router.use("/nurses", nurseRoutes);
